requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
qrcode==7.4.2
Pillow==10.0.0